requirements.txt
python-telegram-bot>=20.0
anthropic
openai
replicate
tweepy>=4.0
python-dotenv
httpx
Pillow
